Service

+++ src/main/java/com/example/hrm_be/repositories/BranchProductRepository.java
@@ -1,6 +1,8 @@
 package com.example.hrm_be.repositories;
 
 import com.example.hrm_be.models.entities.*;
+
+import java.math.BigDecimal;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
@@ -15,5 +17,5 @@ public interface BranchProductRepository
   Optional<BranchProductEntity> findByBranchAndProduct(BranchEntity branch, ProductEntity product);
 
   @Query("SELECT SUM(bp.quantity) FROM BranchProductEntity bp WHERE bp.product.id = :productId")
-  Integer findTotalQuantityForProduct(Long productId);
+  BigDecimal findTotalQuantityForProduct(Long productId);
 }
